perf(css/minifier): reduce visitor
