--- crates/swc_css_minifier/src/compress/mod.rs
@@ -1,1 +1,0 @@
-pub mod at_rule;
